# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wédas - Corporate Recognition Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  employee: 'employee',
  manager: 'manager',
  hr: 'hr',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  department: string;
  position: string;
  role: UserRole;
  active: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  employee: 'employee',
  manager: 'manager',
  hr: 'hr',
} as const;

export interface UserInput {
  name: string;
  email: string;
  password: string;
  department: string;
  position: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  employee: 'employee',
  manager: 'manager',
  hr: 'hr',
} as const;

export interface UserUpdate {
  name?: string;
  department?: string;
  position?: string;
  role?: UserUpdateRole;
  active?: boolean;
}

export interface Category {
  id: number;
  name: string;
  icon: string;
  active: boolean;
}

export interface CategoryInput {
  name: string;
  icon: string;
}

export interface CategoryUpdate {
  name?: string;
  icon?: string;
  active?: boolean;
}

export type RecognitionStatus = typeof RecognitionStatus[keyof typeof RecognitionStatus];


export const RecognitionStatus = {
  active: 'active',
  cancelled: 'cancelled',
} as const;

export interface UserSummary {
  id: number;
  name: string;
  department: string;
  position: string;
}

export interface Recognition {
  id: number;
  senderId: number;
  receiverId: number;
  categoryId: number;
  coins: number;
  message: string;
  status: RecognitionStatus;
  createdAt: string;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  cancelledBy?: number | null;
  sender: UserSummary;
  receiver: UserSummary;
  category: Category;
}

export interface RecognitionInput {
  receiverId: number;
  categoryId: number;
  coins: number;
  /** @maxLength 300 */
  message: string;
}

export interface RankingEntry {
  rank: number;
  user: UserSummary;
  totalCoins: number;
  totalRecognitions: number;
}

export interface CategoryRankingEntry {
  category: Category;
  totalCoins: number;
  totalRecognitions: number;
}

export interface PersonalDashboard {
  availableBalance: number;
  allocatedThisMonth: number;
  receivedThisMonth: number;
  sentThisMonth?: number;
  accumulatedValue: number;
  /** @nullable */
  rankingPosition: number | null;
  recentRecognitions: Recognition[];
}

export interface MonthlyEvolutionEntry {
  month: number;
  year: number;
  totalCoins: number;
  totalRecognitions: number;
}

export interface HrDashboard {
  totalDistributed: number;
  totalRecognitions: number;
  avgPerUser: number;
  topCategory: Category;
  topDepartment: string;
  monthlyEvolution: MonthlyEvolutionEntry[];
  activeUsers?: number;
  engagementRate?: number;
}

export interface Settings {
  id: number;
  coinConversionRate: number;
  monthlyCoinLimit: number;
}

export interface SettingsUpdate {
  coinConversionRate?: number;
  monthlyCoinLimit?: number;
}

export interface UserReportEntry {
  user: User;
  coinsReceived: number;
  coinsSent: number;
  recognitionsReceived: number;
  recognitionsSent: number;
}

export interface FinancialReportEntry {
  user: User;
  totalCoins: number;
  monetaryValue: number;
}

export interface CategoryReportEntry {
  category: Category;
  totalCoins: number;
  totalRecognitions: number;
  percentage: number;
}

export interface MonthlyBalance {
  allocated: number;
  sent: number;
  available: number;
  month: number;
  year: number;
}

export type ListUsersParams = {
search?: string;
department?: string;
};

export type ListRecognitionsParams = {
type?: ListRecognitionsType;
userId?: number;
categoryId?: number;
startDate?: string;
endDate?: string;
limit?: number;
};

export type ListRecognitionsType = typeof ListRecognitionsType[keyof typeof ListRecognitionsType];


export const ListRecognitionsType = {
  sent: 'sent',
  received: 'received',
  all: 'all',
} as const;

export type GetRankingByReceivedParams = {
limit?: number;
};

export type GetRankingBySentParams = {
limit?: number;
};